+++ src/convert/styles/css/vscodeMarkdownStyle.ts
@@ -8,7 +8,6 @@ export const vscodeMarkdownStyle = `
  
  h1, h2, h3, h4, h5, h6 {
    font-weight: 600;
-   margin-top: 24px;
    margin-bottom: 16px;
    line-height: 1.25;
  }
@@ -136,38 +135,7 @@ export const vscodeMarkdownStyle = `
    height: 1px;
    border-bottom: 1px solid;
  }
- 
- h1 {
-   font-size: 2em;
-   margin-top: 0;
-   padding-bottom: 0.3em;
-   border-bottom-width: 1px;
-   border-bottom-style: solid;
- }
- 
- h2 {
-   font-size: 1.5em;
-   padding-bottom: 0.3em;
-   border-bottom-width: 1px;
-   border-bottom-style: solid;
- }
- 
- h3 {
-   font-size: 1.25em;
- }
- 
- h4 {
-   font-size: 1em;
- }
- 
- h5 {
-   font-size: 0.875em;
- }
- 
- h6 {
-   font-size: 0.85em;
- }
- 
+
  table {
    border-collapse: collapse;
    margin-bottom: 0.7em;
@@ -204,32 +172,4 @@ export const vscodeMarkdownStyle = `
  body.wordWrap pre {
    white-space: pre-wrap;
  }
- 
- /** Theming */
- 
- .vscode-high-contrast h1 {
-   border-color: rgb(0, 0, 0);
- }
- 
- .vscode-light th {
-   border-color: rgba(0, 0, 0, 0.69);
- }
- 
- .vscode-dark th {
-   border-color: rgba(255, 255, 255, 0.69);
- }
- 
- .vscode-light h1,
- .vscode-light h2,
- .vscode-light hr,
- .vscode-light td {
-   border-color: rgba(0, 0, 0, 0.18);
- }
- 
- .vscode-dark h1,
- .vscode-dark h2,
- .vscode-dark hr,
- .vscode-dark td {
-   border-color: rgba(255, 255, 255, 0.18);
- }
-`
+ `