+++ src/convert/styles/styleTagBuilder.ts
@@ -2,24 +2,30 @@ import path from "path"
 import * as vscode from 'vscode'
 import { VscodeEnvLanguage, getHomeDirPath, getPaperbackWriterConfiguration, getUri, getVscodeUri, getWorkspaceFolder, showMessage } from "../../vscode-util"
 import { vscodeMarkdownStyle } from "./css/vscodeMarkdownStyle"
-import { codeCss } from "../markdown/renderer/codeCss"
+import { CodeThemeToCssProps, codeThemeToCss } from "../markdown/renderer/codeCss"
 import { remedyCss } from "./css/remedyCss"
 import { FontSetId, getFontFamily } from "./css/fontStyle"
 import { blockquoteCss } from "./css/blockquoteCss"
+import { headerCss } from "./css/headerCss"
 
 var CleanCSS = require('clean-css')
 
 export type StyleTagBuilderProps = {
   editorDocVsUrl: ThemeStyleTagBuilderProps['editorDocVsUrl']
   fontQuerys: GetFontStyleTagsProps['fontQuerys']
+  codeTheme?: CodeThemeToCssProps['theme']
 }
 
 export const styleTagBuilder = ({
   editorDocVsUrl,
-  fontQuerys
+  fontQuerys,
+  codeTheme
 }:StyleTagBuilderProps) => {
 
-  const styleTags = themeStyleTagsBuilder({editorDocVsUrl})
+  const styleTags = themeStyleTagsBuilder({
+    editorDocVsUrl,
+    codeTheme
+  })
   const fontStyleTags = fontStyleTagsBuilder({fontQuerys})
 
   return fontStyleTags + '\n' + styleTags
@@ -27,8 +33,9 @@ export const styleTagBuilder = ({
 
 
 // ------------------------------
-type ThemeStyleTagBuilderProps = {
-  editorDocVsUrl: vscode.Uri
+export type ThemeStyleTagBuilderProps = {
+  editorDocVsUrl: vscode.Uri,
+  codeTheme?: CodeThemeToCssProps['theme']
 }
 type BuildedStyle = string
 
@@ -36,7 +43,8 @@ type BuildedStyle = string
  * テーマに関するスタイルタグを生成する
  */
 export const themeStyleTagsBuilder = ({
-  editorDocVsUrl
+  editorDocVsUrl,
+  codeTheme
 }: ThemeStyleTagBuilderProps): string => {
   const styleTags: string[] = []
   const styleLinks: string[] = []
@@ -50,12 +58,18 @@ export const themeStyleTagsBuilder = ({
     if (includeDefaultStyles) {
       builtInStyles.push(remedyCss)
       builtInStyles.push(vscodeMarkdownStyle)
-      builtInStyles.push(codeCss({}))
+      builtInStyles.push(codeThemeToCss({theme: codeTheme}))
       builtInStyles.push(blockquoteCss())
     }
 
-    // 2. ベースフォントサイズ
+    // 2. header
     builtInStyles.push(`html { font-size: ${PwCnf.style.font.baseSize}px; }`)
+
+    // 3. ベースフォントサイズ
+    builtInStyles.push(headerCss({
+      h1Scale: PwCnf.style.typography.h1HeaderScale
+    }))
+
       
     const minified = new CleanCSS({}).minify(builtInStyles.join('\n')).styles
     styleTags.push(`<style>${minified}</style>`)