+++ CHANGELOG.md
@@ -1,9 +1,29 @@
-# Change Log
+# Changelog
 
-All notable changes to the "paperback-writer" extension will be documented in this file.
+All notable changes to this project will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
-- Initial release
+### Added
+
+- v0.0.3
+  - H1 Header Scale
+  - Syntax-highlighting theming
+  - Syntax Highlighting: Show Line Numbers
+
+### Fixed
+- 
+
+### Changed
+
+- 
+
+### Removed
+
+- 
+
+## [0.0.3] - 2023-03-05
+